updating sim scripts

+++ bin/sim.py
@@ -5,9 +5,29 @@ import subprocess
 import sys
 import os
 
-n = int(sys.argv[1])
-k = int(sys.argv[2])
-analysis = sys.argv[3]
+from optparse import OptionParser
+
+parser = OptionParser()
+parser.add_option("-n", "--nreps", dest="n",
+                  help="number of replicates", metavar="N")
+parser.add_option("-k", "--ncpu", dest="k",
+                  help="number of cpus", metavar="K")
+parser.add_option("-m", "--model", dest="model",
+                  help="morphological model to use", default="none", metavar="MODEL")
+parser.add_option("-t", "--tree",
+                  action="store_true", dest="tree", default=False,
+                  help="use the full FBD tree model")
+
+(options, args) = parser.parse_args()
+
+models = ['asym','mk','none']
+
+if options.model not in models:
+	raise Exception("Morphological model must be one of: "+str(models))
+
+if not os.path.isfile("params.r"):
+	raise Exception("params.r file not found")
+
 
 os.system('rm -rf sims')
 os.system('mkdir sims')
@@ -17,19 +37,36 @@ simdir = os.path.abspath(os.getcwd())
 
 def simulate(i):
 	os.chdir(simdir+'/sims')
-	ext = ("%0"+str(len(str(n)))+"d") % (i + 1)
+	ext = ("%0"+str(len(options.n))+"d") % (i + 1)
 	os.system('mkdir sim'+ext)
 	os.chdir('sim'+ext)
 
 	os.system(bindir+'/simfbd.r ../../params.r')
-	if analysis == 'morpho':
-		os.system('rb '+bindir+'/simnex.rev')
-	os.system('rb '+bindir+'/infer-'+analysis+'.rev')
-	os.system(bindir+'/summarize.r '+analysis)
 
-pool = mp.Pool(processes=k)
+	if options.model == 'asym':
+		os.system('rb '+bindir+'/sim-asym.rev')
+		os.system('cp fbd.taxa diagnosed.taxa')
+		os.system('cp asym.nex diagnosed.nex')
+	elif options.model == 'mk':
+		os.system('rb '+bindir+'/sim-mk.rev')
+		os.system(bindir+'/diagnose.py mk.fa sa.taxa')
+	else:
+		os.system('cp fbd.taxa diagnosed.taxa')
 
-for i in range(n):
+	if options.tree:
+		if options.model == 'none':
+			os.system('rb '+bindir+'/infer-tree.rev')
+			os.system(bindir+'/summarize.r')
+		else:
+			os.system('rb '+bindir+'/infer-mk.rev')
+			os.system(bindir+'/summarize.r 1')
+	else:
+		os.system('rb '+bindir+'/infer-range.rev')
+		os.system(bindir+'/summarize.r')
+
+pool = mp.Pool(processes=int(options.k))
+
+for i in range(int(options.n)):
 	pool.apply_async(simulate, args=(i,))
 
 pool.close()
